pyproject.toml
[project]
name = "tg-admin-caller-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-telegram-bot>=22.3",
    "ruff>=0.12.10",
]
